Added basic functionality to exclude LAS points based on classification
Added argument to the points2grid script to filter out LAS points based on its classification.

Added a method to the interpolation class to set the exclude classification called "setLasExcludeClassification". Modified the interpolation method to exclude points if need be.

Edited the lasfile header to have a method to return the point classification.